backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
pymorphy3==2.0.6
httpx==0.27.2
playwright==1.49.1
python-multipart==0.0.20
aiofiles==24.1.0
python-docx==1.1.2
pdfminer.six==20240706
pytesseract==0.3.13
Pillow==11.0.0
greenlet==3.1.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
